server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[daily,elevenlabs,openai,silero,google]==0.0.71
aws_sdk_bedrock_runtime
runner
loguru
strands-agents 
strands-agents-tools